A challlenge based on Numpy and pandas
Consider a datasets that is a pair of sentences as follows

Who are you? , Helen.
Where Are you?, I am at home.

Now consider the machine only knows these 4 sentences (The machine is a baby👶 who has only heard these 4 sentences ever in its life and has ridiculously good memory).
It's vocabulary should contain only

  0       1      2       3          4        5    6    7    8
['who' 'are' 'you' 'helen' 'where' 'i' 'am' 'at' 'home'] (vocabulary length - 9)

Now the objective is converting each sentence in such a way, the uniqueness of the words are not lost as well as its order.
For this we imagine an 1-dimensional numpy array to represent each word. This numpy array has same length as the length of vocabulary (here, 9) such that
   > value 0 if the word at that index of vocabulary list is not the word the numpy array represents
   > value 1 if the word at that index of vocabulary list is the word the numpy array represents

example-

who  - 1,0,0,0,0,0,0,0,0
are  - 0,1,0,0,0,0,0,0,0
you - 0,0,1,0,0,0,0,0,0

Therefore, the sentence as a whole becomes a 2-dimensional array as 
[ [1,0,0,0,0,0,0,0,0],
  [0,1,0,0,0,0,0,0,0],
  [0,0,1,0,0,0,0,0,0],]

Your mission, should you choose to accept it, is to make a google colab notebook which
>  Provides an upload option to upload the dataset(CSV file) and
> Step by Step conversion of the pair of sentences into pairs of 2-dimensional numpy arrays as explained above. (include text cells describing what each part of your code does, (to prevent accidental miracles 😜))
>  print the vocabulary list and the pairs of 2-dimensional arrays of each row in input.

*NOTES*
>  The vocabulary list is fixed throughout a single data set and not just 1 sentence. Hence the shape of each sentence would be 
   number of words in sentence, number of words in vocabulary
   So you should consider the full data set when making the vocabulary list.
> You may order the vocabulary list your own way but the same list should be used to find the 2-dimensional arrays throughout a single data set.
> you can ignore and remove punctuation (for now), they are just annoying.
> Sample input data can be found here -> http://dravog.me/data/input.csv
   Do remember to ignore the header row.
> Sample output can be found here -> http://dravog.me/data/output.csv

###########################3
I did the coding part in colloab so some part of the code works only in it.